deduplicate the code
